update user height data type

+++ app/src/main/java/com/nutriomatic/app/data/remote/api/response/ProfileResponse.kt
@@ -83,5 +83,5 @@ data class User(
 	val profpic: String,
 
 	@field:SerializedName("height")
-	val height: Int
+	val height: Double
 )
